Does way too much in one commit.
* add a findAndReplaceFiles glob options. When set, any matches will 
have the logical path for each fingerprinted-file-so-far replaced with
the fingerprinted path. This allows the user to carefully order  
things so that files that reference other files can be sure to come after
them in the process. non-fingerprinted files (like HTML templates) can
have the search and replace performed on them as well.
* add a keepOriginalFiles option. [true] (No effect on current use)
* pretty prints dist/assets.json

Example config: 

    assetFingerprint:
      options:
        findAndReplaceFiles: ["dist/**/*.js", "dist/**/*.css", "dist/**/*.html", "dist/**/*.xml"]
        keepOriginalFiles: false