Avoid some false positives when prefixing root-relative paths with CDN
In a case like `../originalName.png`, you would end up with
`..https://cdn.domain.com/hashedName.png`. In this scenario, you would
actually want to keep the relative path still `../hashedName.png`, but
with the fingerprinted filename.

Although we don't have a language parser to ensure we are looking at a
root-relative name, we can at least take try to be a bit more cautious
and look out for common boundaries.